// frontend/src/api/generated/models/character-create.ts
/* tslint:disable */
 
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * キャラクター作成スキーマ
 * @export
 * @interface CharacterCreate
 */
export interface CharacterCreate {
    /**
     * キャラクター名
     * @type {string}
     * @memberof CharacterCreate
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof CharacterCreate
     */
    'description'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CharacterCreate
     */
    'appearance'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CharacterCreate
     */
    'personality'?: string | null;
    /**
     * 現在地
     * @type {string}
     * @memberof CharacterCreate
     */
    'location'?: string;
}